Fixed a few bugs
Adding an extensions didn't update or threw an exception.
Renaming a category sometimes threw an exception.
Fixed a bug with the context menu items including a backslash at the end.

+++ SortingEngine/SortEngine.cs
@@ -9,47 +9,45 @@ namespace SortingEngine
 {
     public class Sorter
     {
-        public List<FileItem> FileList = new List<FileItem>();
-
+        string ActiveDirectory { get; set; } //Application generated folder
         public string CurrentDirectory { get; set; } //Choosen directory
-        private string ActiveDirectory { get; set; } //Application generated folder
 
-        private void MoveFiles(List<FileItem> filelist, string ActDir)
+        List<FileItem> FileList = new List<FileItem>();
+
+        void MoveFiles(List<FileItem> filelist, string ActDir)
         {
             foreach (FileItem file in filelist)
             {
                 try
-                {//Move file if no file with same name exists.
+                {
+                    //Move file if no file with same name exists.
                     if (!File.Exists(ActDir + file.NameWithExtension))
                         File.Move(file.FullPath, ActDir + file.NameWithExtension);
+
                     else
-                    {//Creates a loop that increases int i to a maximum of file.count in active directory & adds i to filename to avoid duplicate exception. Ends with return to avoid multiple executions.
+                    {
+                        //Loops an int++ until a valid name has been generated.
                         for (int i = 0; i <= Directory.GetFiles(ActDir).Count(); i++)
                         {
                             if (!File.Exists(ActDir + file.Filename + $"({i})" + file.Extension))
                             {
                                 File.Move(file.FullPath, ActDir + file.Filename + $"({i})" + file.Extension);
-                                return;
+                                break;
                             }
 
                         }
                     }
                 }
-                catch (Exception E)
-                {//incase of unexpected exception.
-                    MessageBox.Show(E.Message);
-                    return; 
-                }
+                catch (Exception E) { MessageBox.Show(E.Message); }
             }
         }
 
-        private void DeleteFolder()
+        void DeleteFolder()
         {
-            if (Directory.Exists(ActiveDirectory))
-                if (Directory.GetFiles(ActiveDirectory).Count() == 0)
-                    Directory.Delete(ActiveDirectory);
-                else return;
-            else return;
+            if (Directory.Exists(ActiveDirectory) &&
+                Directory.GetFiles(ActiveDirectory).Count() == 0 &&
+                Directory.GetDirectories(ActiveDirectory).Count() == 0)
+                Directory.Delete(ActiveDirectory);
         }
 
         public void Sort(Category Cat)
@@ -64,10 +62,21 @@ namespace SortingEngine
             if (TempList.Count() > 0)
             {
                 foreach (string S in TempList)
+                {
+                    #region Failsafe since extensions like .doc and .docx create duplicate entries
+                    bool Exists = false;
+
+                    foreach (FileItem F in FileList)
+                        if (S == F.FullPath) { Exists = true; break; }
+
+                    if (Exists) continue;
+                    #endregion
+
                     FileList.Add(new FileItem(S));
+                }
 
                 if (!Directory.Exists(ActiveDirectory))
-                    Directory.CreateDirectory(ActiveDirectory); 
+                    Directory.CreateDirectory(ActiveDirectory);
 
                 MoveFiles(FileList, ActiveDirectory);
             }
@@ -80,12 +89,12 @@ namespace SortingEngine
 
     public class UnSorter
     {
-        public List<string> FileList = new List<string>();
-
+        string ActiveDirectory { get; set; }
         public string CurrentDirectory { get; set; }
-        private string ActiveDirectory { get; set; }
 
-        private void MoveFiles(List<string> filelist)
+        List<string> FileList = new List<string>();
+
+        void MoveFiles(List<string> filelist)
         {
             foreach (string file in filelist)
             {
@@ -93,17 +102,15 @@ namespace SortingEngine
                 {
                     File.Move(file, CurrentDirectory + Path.GetFileName(file));
                 }
-                catch (Exception) { return; }
+                catch (Exception E) { MessageBox.Show(E.Message); }
             }
         }
 
-        private void DeleteFolder()
+        void DeleteFolder()
         {
-            if (Directory.Exists(ActiveDirectory))
-                if (Directory.GetFiles(ActiveDirectory).Count() == 0)
-                    Directory.Delete(ActiveDirectory);
-                else return;
-            else return;
+            if (Directory.GetFiles(ActiveDirectory).Count() == 0 &&
+                Directory.GetDirectories(ActiveDirectory).Count() == 0)
+                Directory.Delete(ActiveDirectory);
         }
 
         public void UnSort(Category Cat)
@@ -119,7 +126,6 @@ namespace SortingEngine
 
                 DeleteFolder();
             }
-            else return;
 
             FileList.Clear();
         }
